--- Structs/record.go
@@ -1,7 +1,0 @@
-package Structs
-
-type Record struct {
-	T    bool
-	Num  int
-	Body string
-}
