+++ pkg/api/questions.go
@@ -1,7 +1,7 @@
 package api
 
 import (
-	"dz/Structs"
+	"dz/pkg/models"
 	"encoding/json"
 	"net/http"
 )
@@ -15,14 +15,12 @@ func (api *api) QuestionHandle(w http.ResponseWriter, r *http.Request) {
 		}
 
 	case http.MethodPost:
-		api.s.Lock()
-		var m []Structs.Record
+
+		var m []models.Record
 		err := json.NewDecoder(r.Body).Decode(&m)
 		if err != nil {
-			api.s.Unlock()
 			return
 		}
 		api.s.AddFromJson(m)
-		api.s.Unlock()
 	}
 }
