+++ pkg/models/Record.go
@@ -1,0 +1,7 @@
+package models
+
+type Record struct {
+	IsQuestion bool
+	Num        int
+	Body       string
+}
